暗黙のキャストへの警告の対応2

+++ src/Grid/Field.cpp
@@ -32,31 +32,31 @@ void Field::setBoundaryConditionPhi(void) {
 }
 
 void Field::setNeumannPhi(const AXIS axis, const AXIS_SIDE low_or_up) {
-    const int max_index = phi.shape()[ Utils::getAxisIndex(axis) ];
+    const size_t max_index = phi.shape()[ Utils::getAxisIndex(axis) ];
 
-    const int boundary_index = (low_or_up == AXIS_SIDE::low) ? 1 : (max_index - 2);
-    const int neighbor_element_index = (low_or_up == AXIS_SIDE::low) ? 2 : (max_index - 3);
+    const size_t boundary_index = (low_or_up == AXIS_SIDE::low) ? 1 : (max_index - 2);
+    const size_t neighbor_element_index = (low_or_up == AXIS_SIDE::low) ? 2 : (max_index - 3);
 
-    const int cx_with_glue = phi.shape()[0];
-    const int cy_with_glue = phi.shape()[1];
-    const int cz_with_glue = phi.shape()[2];
+    const size_t cx_with_glue = phi.shape()[0];
+    const size_t cy_with_glue = phi.shape()[1];
+    const size_t cz_with_glue = phi.shape()[2];
 
     if ( Environment::isOnEdge(axis, low_or_up) ) {
         if (axis == AXIS::x) {
-            for(int k = 1; k < cz_with_glue - 1; ++k){
-                for(int j = 1; j < cy_with_glue - 1; ++j){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
+                for(size_t j = 1; j < cy_with_glue - 1; ++j){
                     phi[boundary_index][j][k] = phi[neighbor_element_index][j][k];
                 }
             }
         } else if (axis == AXIS::y) {
-            for(int k = 1; k < cz_with_glue - 1; ++k){
-                for(int i = 1; i < cx_with_glue - 1; ++i){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
+                for(size_t i = 1; i < cx_with_glue - 1; ++i){
                     phi[i][boundary_index][k] = phi[i][neighbor_element_index][k];
                 }
             }
         } else {
-            for(int j = 1; j < cy_with_glue - 1; ++j){
-                for(int i = 1; i < cx_with_glue - 1; ++i){
+            for(size_t j = 1; j < cy_with_glue - 1; ++j){
+                for(size_t i = 1; i < cx_with_glue - 1; ++i){
                     phi[i][j][boundary_index] = phi[i][j][neighbor_element_index];
                 }
             }
@@ -65,29 +65,29 @@ void Field::setNeumannPhi(const AXIS axis, const AXIS_SIDE low_or_up) {
 }
 
 void Field::setDirichletPhi(const AXIS axis, const AXIS_SIDE low_or_up) {
-    const int max_index = phi.shape()[ Utils::getAxisIndex(axis) ];
-    const int boundary_index = (low_or_up == AXIS_SIDE::low) ? 1 : (max_index - 2);
+    const size_t max_index = phi.shape()[ Utils::getAxisIndex(axis) ];
+    const size_t boundary_index = (low_or_up == AXIS_SIDE::low) ? 1 : (max_index - 2);
 
-    const int cx_with_glue = phi.shape()[0];
-    const int cy_with_glue = phi.shape()[1];
-    const int cz_with_glue = phi.shape()[2];
+    const size_t cx_with_glue = phi.shape()[0];
+    const size_t cy_with_glue = phi.shape()[1];
+    const size_t cz_with_glue = phi.shape()[2];
 
     if ( Environment::isOnEdge(axis, low_or_up) ) {
         if (axis == AXIS::x) {
-            for(int k = 1; k < cz_with_glue - 1; ++k){
-                for(int j = 1; j < cy_with_glue - 1; ++j){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
+                for(size_t j = 1; j < cy_with_glue - 1; ++j){
                     phi[boundary_index][j][k] = 0.0;
                 }
             }
         } else if (axis == AXIS::y) {
-            for(int k = 1; k < cz_with_glue - 1; ++k){
-                for(int i = 1; i < cx_with_glue - 1; ++i){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
+                for(size_t i = 1; i < cx_with_glue - 1; ++i){
                     phi[i][boundary_index][k] = 0.0;
                 }
             }
         } else {
-            for(int j = 1; j < cy_with_glue - 1; ++j){
-                for(int i = 1; i < cx_with_glue - 1; ++i){
+            for(size_t j = 1; j < cy_with_glue - 1; ++j){
+                for(size_t i = 1; i < cx_with_glue - 1; ++i){
                     phi[i][j][boundary_index] = 0.0;
                 }
             }
@@ -97,15 +97,15 @@ void Field::setDirichletPhi(const AXIS axis, const AXIS_SIDE low_or_up) {
 
 //! FDTD法ベースで電場を更新する
 void Field::updateEfieldFDTD(const double dx, const double dt) {
-    const int cx_with_glue = ex.shape()[0] + 1; // nx + 2
-    const int cy_with_glue = ey.shape()[1] + 1;
-    const int cz_with_glue = ez.shape()[2] + 1;
+    const size_t cx_with_glue = ex.shape()[0] + 1; // nx + 2
+    const size_t cy_with_glue = ey.shape()[1] + 1;
+    const size_t cz_with_glue = ez.shape()[2] + 1;
     const double dt_per_eps0 = dt / Normalizer::eps0;
     const double dt_per_mu0_eps0_dx = dt_per_eps0 / (Normalizer::mu0 * dx);
 
-    for(int i = 1; i < cx_with_glue - 1; ++i){
-        for(int j = 1; j < cy_with_glue - 1; ++j){
-            for(int k = 1; k < cz_with_glue - 1; ++k){
+    for(size_t i = 1; i < cx_with_glue - 1; ++i){
+        for(size_t j = 1; j < cy_with_glue - 1; ++j){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
                 //! 各方向には1つ少ないのでcx-1まで
                 if(i < cx_with_glue - 2) {
                     ex[i][j][k] = ex[i][j][k] - jx[i][j][k] * dt_per_eps0 +
@@ -138,13 +138,13 @@ void Field::updateEfieldFDTD(const double dx, const double dt) {
 }
 
 void Field::setDampingBoundaryOnEfield(void) {
-    const int cx_with_glue = ex.shape()[0] + 1;
-    const int cy_with_glue = ey.shape()[1] + 1;
-    const int cz_with_glue = ez.shape()[2] + 1;
+    const size_t cx_with_glue = ex.shape()[0] + 1;
+    const size_t cy_with_glue = ey.shape()[1] + 1;
+    const size_t cz_with_glue = ez.shape()[2] + 1;
 
     if (!Environment::isNotBoundary(AXIS::x, AXIS_SIDE::low)) {
-        for(int k = 0; k < cz_with_glue; ++k){
-            for(int j = 0; j < cy_with_glue; ++j){
+        for(size_t k = 0; k < cz_with_glue; ++k){
+            for(size_t j = 0; j < cy_with_glue; ++j){
                 ex[0][j][k] = 0.0; // glue cell
                 ex[1][j][k] = 0.0;
             }
@@ -152,8 +152,8 @@ void Field::setDampingBoundaryOnEfield(void) {
     }
 
     if (!Environment::isNotBoundary(AXIS::x, AXIS_SIDE::up)) {
-        for(int k = 0; k < cz_with_glue; ++k){
-            for(int j = 0; j < cy_with_glue; ++j){
+        for(size_t k = 0; k < cz_with_glue; ++k){
+            for(size_t j = 0; j < cy_with_glue; ++j){
                 //! ex は x方向に 1 小さいので、cx_with_glue - 2 が glue cell になる
                 ex[cx_with_glue - 2][j][k] = 0.0; // glue cell
                 ex[cx_with_glue - 3][j][k] = 0.0;
@@ -162,8 +162,8 @@ void Field::setDampingBoundaryOnEfield(void) {
     }
 
     if (!Environment::isNotBoundary(AXIS::y, AXIS_SIDE::low)) {
-        for(int i = 0; i < cx_with_glue; ++i){
-            for(int k = 0; k < cz_with_glue; ++k){
+        for(size_t i = 0; i < cx_with_glue; ++i){
+            for(size_t k = 0; k < cz_with_glue; ++k){
                 ey[i][0][k] = 0.0; // glue cell
                 ey[i][1][k] = 0.0;
             }
@@ -171,8 +171,8 @@ void Field::setDampingBoundaryOnEfield(void) {
     }
 
     if (!Environment::isNotBoundary(AXIS::y, AXIS_SIDE::up)) {
-        for(int i = 0; i < cx_with_glue; ++i){
-            for(int k = 0; k < cz_with_glue; ++k){
+        for(size_t i = 0; i < cx_with_glue; ++i){
+            for(size_t k = 0; k < cz_with_glue; ++k){
                 ey[i][cy_with_glue - 2][k] = 0.0; // glue cell
                 ey[i][cy_with_glue - 3][k] = 0.0;
             }
@@ -180,8 +180,8 @@ void Field::setDampingBoundaryOnEfield(void) {
     }
 
         if (!Environment::isNotBoundary(AXIS::z, AXIS_SIDE::low)) {
-        for(int i = 0; i < cx_with_glue; ++i){
-            for(int j = 0; j < cy_with_glue; ++j){
+        for(size_t i = 0; i < cx_with_glue; ++i){
+            for(size_t j = 0; j < cy_with_glue; ++j){
                 ez[i][j][0] = 0.0; // glue cell
                 ez[i][j][1] = 0.0;
             }
@@ -189,8 +189,8 @@ void Field::setDampingBoundaryOnEfield(void) {
     }
 
     if (!Environment::isNotBoundary(AXIS::z, AXIS_SIDE::up)) {
-        for(int i = 0; i < cx_with_glue; ++i){
-            for(int j = 0; j < cy_with_glue; ++j){
+        for(size_t i = 0; i < cx_with_glue; ++i){
+            for(size_t j = 0; j < cy_with_glue; ++j){
                 ez[i][j][cz_with_glue - 2] = 0.0; // glue cell
                 ez[i][j][cz_with_glue - 3] = 0.0;
             }
@@ -259,9 +259,9 @@ void Field::initializeCurrent(const double dt) {
     Utils::initialize3DArray(jy);
     Utils::initialize3DArray(jz);
 
-    const int cx_with_glue = jx.shape()[0] + 1; // Edge 要素は各方向に1少ないので +1 する
-    const int cy_with_glue = jy.shape()[1] + 1;
-    const int cz_with_glue = jz.shape()[2] + 1;
+    const size_t cx_with_glue = jx.shape()[0] + 1; // Edge 要素は各方向に1少ないので +1 する
+    const size_t cy_with_glue = jy.shape()[1] + 1;
+    const size_t cz_with_glue = jz.shape()[2] + 1;
 
     //! 背景電流などがある場合にはここで設定する
 
@@ -271,23 +271,23 @@ void Field::initializeCurrent(const double dt) {
     // cout << "[NOTICE] " << 1.0 / Normalizer::normalizeFrequency(real_freq) << " step で 1周期です." << endl;
     const double freq = 2.0 * M_PI * Normalizer::normalizeFrequency(real_freq); // Hz
     const double J0 = 10.0;
-    const int half_x = cx_with_glue / 2;
-    const int half_y = cy_with_glue / 2;
-    for (int k = 0; k < cz_with_glue - 1; ++k) {
+    const size_t half_x = cx_with_glue / 2;
+    const size_t half_y = cy_with_glue / 2;
+    for (size_t k = 0; k < cz_with_glue - 1; ++k) {
         jz[half_x][half_y][k] = J0 * std::sin(freq * now);
     }
 }
 
 double Field::getEfieldEnergy(void) const {
-    const int cx_with_glue = ey.shape()[0];
-    const int cy_with_glue = ex.shape()[1];
-    const int cz_with_glue = ex.shape()[2];
+    const size_t cx_with_glue = ey.shape()[0];
+    const size_t cy_with_glue = ex.shape()[1];
+    const size_t cz_with_glue = ex.shape()[2];
 
     double energy = 0.0;
 
-    for(int i = 1; i < cx_with_glue - 1; ++i){
-        for(int j = 1; j < cy_with_glue - 1; ++j){
-            for(int k = 1; k < cz_with_glue - 1; ++k){
+    for(size_t i = 1; i < cx_with_glue - 1; ++i){
+        for(size_t j = 1; j < cy_with_glue - 1; ++j){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
                 //! 各点のエネルギーを計算する(Yee格子内のエネルギーの計算方法は?)
                 if (i < cx_with_glue - 2) energy += pow(ex[i][j][k], 2);
                 if (j < cy_with_glue - 2) energy += pow(ey[i][j][k], 2);
@@ -300,15 +300,15 @@ double Field::getEfieldEnergy(void) const {
 }
 
 double Field::getBfieldEnergy(void) const {
-    const int cx_with_glue = bx.shape()[0];
-    const int cy_with_glue = by.shape()[1];
-    const int cz_with_glue = bz.shape()[2];
+    const size_t cx_with_glue = bx.shape()[0];
+    const size_t cy_with_glue = by.shape()[1];
+    const size_t cz_with_glue = bz.shape()[2];
 
     double energy = 0.0;
 
-    for(int i = 1; i < cx_with_glue - 1; ++i){
-        for(int j = 1; j < cy_with_glue - 1; ++j){
-            for(int k = 1; k < cz_with_glue - 1; ++k){
+    for(size_t i = 1; i < cx_with_glue - 1; ++i){
+        for(size_t j = 1; j < cy_with_glue - 1; ++j){
+            for(size_t k = 1; k < cz_with_glue - 1; ++k){
                 //! 各点のエネルギーを計算する(Yee格子内のエネルギーの計算方法は?)
                 if ( (j < cy_with_glue - 2) && (k < cz_with_glue - 2) ) energy += pow(bx[i][j][k], 2);
                 if ( (i < cx_with_glue - 2) && (k < cz_with_glue - 2) ) energy += pow(by[i][j][k], 2);
@@ -324,13 +324,13 @@ void Field::checkChargeConservation(const RhoArray& old_rho, const double dt, co
     double residual1 = 0.0, residual2 = 0.0;
     double j1 = 0.0, j2 = 0.0, j3 = 0.0;
     double residual = 0.0;
-    const int cx_with_glue = rho[0].shape()[0];
-    const int cy_with_glue = rho[0].shape()[1];
-    const int cz_with_glue = rho[0].shape()[2];
+    const size_t cx_with_glue = rho[0].shape()[0];
+    const size_t cy_with_glue = rho[0].shape()[1];
+    const size_t cz_with_glue = rho[0].shape()[2];
 
-    for(int i = 1; i < cx_with_glue - 1; ++i) {
-        for(int j = 1; j < cy_with_glue - 1; ++j) {
-            for(int k = 1; k < cz_with_glue - 1; ++k) {
+    for(size_t i = 1; i < cx_with_glue - 1; ++i) {
+        for(size_t j = 1; j < cy_with_glue - 1; ++j) {
+            for(size_t k = 1; k < cz_with_glue - 1; ++k) {
                 residual1 = 0.0, residual2 = 0.0;
                 j1 = 0.0, j2 = 0.0, j3 = 0.0;
